# Remove unneeded overwrite and call to anyway inherited __hash__ method