--- a/src/zope/interface/declarations.py
+++ b/src/zope/interface/declarations.py
@@ -223,9 +223,6 @@ class Implements(Declaration):
         # This spelling works under Python3, which doesn't have cmp().
         return (n1 > n2) - (n1 < n2)
 
-    def __hash__(self):
-        return Declaration.__hash__(self)
-
     # We want equality to be based on identity. However, we can't actually
     # implement __eq__/__ne__ to do this because sometimes we get wrapped in a proxy.
     # We need to let the proxy types implement these methods so they can handle unwrapping