Remove unneeded overwrite and call to anyway inherited `__hash__` method  from `.declarations.Implements` class. 

Watching a reindex index process in ZCatalog with Py-Spy after 10k samples the TotalTime for `.adapter._lookup`  was reduced from 27.5s to 18.8s (~1.5x faster) by taking almost the same OwnTime (before 15.7s/after 13.18s). Before the removed `__hash__` had an OwnTime of 7.76s.

Overall reindex index time  shrunk from 369s to 293s (1.26x faster).

I wonder why this method was necessary? Or is it still for some edge case? 
I am sure it was not placed there w/o a reason.