@thet 2020-01-28
@jamadden added it here: https://github.com/zopefoundation/zope.interface/commit/bbf1a3d4792bf5b0a3969746521b27d2a2b4bf06#diff-135050e0178a707f1f3c1fd503523f70R176