@jamadden 2020-01-28
Prior to some recent commit, there wasn't a `__hash__` that could correctly be inherited. Without defining it here, you get `TypeError`:
```python traceback
  File "//src/zope/interface/interface.py", line 669, in _wire
    from zope.interface.interfaces import IAttribute
  File "//src/zope/interface/interfaces.py", line 770, in <module>
    class ObjectEvent(object):
  File "//src/zope/interface/declarations.py", line 385, in __call__
    classImplements(ob, *self.interfaces)
  File "//src/zope/interface/declarations.py", line 323, in classImplements
    spec = implementedBy(cls)
  File "//src/zope/interface/declarations.py", line 274, in implementedByFallback
    spec = Implements.named(spec_name, *[implementedBy(c) for c in bases])
  File "//src/zope/interface/declarations.py", line 274, in <listcomp>
    spec = Implements.named(spec_name, *[implementedBy(c) for c in bases])
  File "//src/zope/interface/declarations.py", line 274, in implementedByFallback
    spec = Implements.named(spec_name, *[implementedBy(c) for c in bases])
  File "//src/zope/interface/declarations.py", line 142, in named
    inst.__init__(*interfaces)
  File "//src/zope/interface/declarations.py", line 64, in __init__
    Specification.__init__(self, _normalizeargs(interfaces))
  File "//src/zope/interface/interface.py", line 190, in __init__
    self.__bases__ = tuple(bases)
  File "//src/zope/interface/interface.py", line 214, in __setBases
    self.changed(self)
  File "//src/zope/interface/declarations.py", line 67, in changed
    Specification.changed(self, originally_changed)
  File "//src/zope/interface/interface.py", line 233, in changed
    ancestors = ro(self)
  File "//src/zope/interface/ro.py", line 64, in ro
    return _mergeOrderings([_flatten(object)])
  File "//src/zope/interface/ro.py", line 41, in _mergeOrderings
    if o not in seen:
TypeError: unhashable type: 'Implements'
```

It's worth spending a few minutes tracking down the exact change that made that unnecessary and verifying that the hashes haven't changed/are still consistent with comparisons and equality. Please hold ☎️