@jamadden 2020-01-28
Prior to 4.3.2, `Implements` defined `__eq__`, which meant that it had to define `__hash__`, or you get the `TypeError` mentioned above. `__eq__` was defined as identity (`self is other`) and we needed the matching identity `__hash__`, `object.__hash__`. So `def __hash__(self): return Declaration.__hash__(self)` was the way that was spelled. (Today I'd see if just `__hash__ = object.__hash__` works; that should also eliminate this speed penalty.)

That was all made moot in 4.3.2 with https://github.com/zopefoundation/zope.interface/issues/55 which removed the definition of `__eq__` entirely. Without an explicit `__eq__` defined, we don't have to have an explicit `__hash__` defined to avoid the `TypeError`, we inherit object's version by default.